Latest working
This works! This displays the first order, shows payment option buttons, and fetches orders continuously. There is a problem with the payment buttons showing up sometimes. I had to open a different app and come back to see the updated buttons on screen. Not sure how this works or why it's happening- maybe something with the refresh rate idk why that would be. not sure really